requirements.txt
djangorestframework
Django~=3.1.2